🚚 move `DeploymentCard` component into its own file
